[WIP]reset flip-flop
First spice netlist was made based on extracted sff1_x4 one and this
netlist was used to make layout. Layout seems to be OK now, verification
has been done on extracted netlist by a spice test bench. LVS is failing
though as construction seems not to be recognized properly.

This is work-in-progress and not meant to be upstreamed (yet).